Weekly loop refactor
